Extract promo subscription in own module
