+++ lib/sanbase/billing/subscription.ex
@@ -17,14 +17,6 @@ defmodule Sanbase.Billing.Subscription do
 
   require Logger
 
-  @promo_trial_period_days 14
-  @doc """
-  Enterprise plans from API, Sanbase and Sheets products are used for #{@promo_trial_period_days} days
-  promotional free trial of our services.
-  """
-  @promo_plans [5, 14, 24]
-  @promo_metadata %{"promo_event" => "devcon2019"}
-
   @percent_discount_1000_san 20
   @percent_discount_200_san 4
   @generic_error_message """
@@ -145,32 +137,6 @@ defmodule Sanbase.Billing.Subscription do
 
       {:error, reason} ->
         {:error, reason}
-    end
-  end
-
-  @doc """
-  Attach a subscription with a free trial of #{@promo_trial_period_days} days for all highest priced plans for all
-  products to the user. It doesn't require a credit card.
-  """
-  def promo_subscription(%User{stripe_customer_id: stripe_customer_id} = user)
-      when is_binary(stripe_customer_id) do
-    promo_subscribe(user)
-  end
-
-  def promo_subscription(%User{} = user) do
-    create_or_update_stripe_customer(user)
-    |> case do
-      {:ok, user} ->
-        promo_subscribe(user)
-
-      {:error, error} ->
-        Logger.error(
-          "Error creating promotional subscription for user: #{inspect(user)}, reason: #{
-            inspect(error)
-          }"
-        )
-
-        {:error, @generic_error_message}
     end
   end
 
@@ -309,11 +275,10 @@ defmodule Sanbase.Billing.Subscription do
 
   def plan_name(subscription), do: subscription.plan.name
 
-  # Private functions
-  defp create_or_update_stripe_customer(_user, params \\ %{})
+  def create_or_update_stripe_customer(_user, params \\ %{})
 
-  defp create_or_update_stripe_customer(%User{stripe_customer_id: stripe_id} = user, params)
-       when is_nil(stripe_id) do
+  def create_or_update_stripe_customer(%User{stripe_customer_id: stripe_id} = user, params)
+      when is_nil(stripe_id) do
     StripeApi.create_customer(user, params)
     |> case do
       {:ok, stripe_customer} ->
@@ -326,8 +291,8 @@ defmodule Sanbase.Billing.Subscription do
     end
   end
 
-  defp create_or_update_stripe_customer(%User{stripe_customer_id: stripe_id} = user, params)
-       when is_binary(stripe_id) do
+  def create_or_update_stripe_customer(%User{stripe_customer_id: stripe_id} = user, params)
+      when is_binary(stripe_id) do
     StripeApi.update_customer(user, params)
     |> case do
       {:ok, _} ->
@@ -337,6 +302,8 @@ defmodule Sanbase.Billing.Subscription do
         {:error, reason}
     end
   end
+
+  # Private functions
 
   defp create_stripe_subscription(user, plan) do
     subscription = %{
@@ -397,43 +364,5 @@ defmodule Sanbase.Billing.Subscription do
       {:ok, %Decimal{} = balance} -> balance |> Decimal.to_float()
       _ -> 0
     end
-  end
-
-  defp promo_subscribe(user) do
-    from(p in Plan, where: p.id in ^@promo_plans)
-    |> Repo.all()
-    |> Enum.map(&promo_subscribe(user, &1))
-    |> Enum.filter(&match?({:error, _}, &1))
-    |> case do
-      [] ->
-        {:ok, user_subscriptions(user)}
-
-      errors ->
-        Logger.error(
-          "Error creating promotional subscription for user: #{inspect(user)}, reason: #{
-            inspect(errors)
-          }"
-        )
-
-        {:error, @generic_error_message}
-    end
-  end
-
-  defp promo_subscribe(user, plan) do
-    with {:ok, stripe_subscription} <-
-           promotional_subsciption_data(user, plan) |> StripeApi.create_subscription(),
-         {:ok, subscription} <- create_subscription_db(stripe_subscription, user, plan) do
-      {:ok, subscription}
-    end
-  end
-
-  defp promotional_subsciption_data(user, plan) do
-    %{
-      customer: user.stripe_customer_id,
-      items: [%{plan: plan.stripe_id}],
-      trial_period_days: @promo_trial_period_days,
-      metadata: @promo_metadata,
-      cancel_at: Timex.shift(Timex.now(), days: @promo_trial_period_days) |> DateTime.to_unix()
-    }
   end
 end
